Untested WIP: Written fxns for setting the indexName on dappsmith and passing in the cdnURL.  Added handler for CodePipeline integration which updates the Cloudfront distro, updates the S3 website config, and sends the owner a confirmation email

+++ index.js
@@ -1,5 +1,6 @@
 'use strict';
 const api = require('./api');
+const cleanup = require('./services/cleanup');
 const validate = require('./validate');
 
 exports.handler = async (event) => {
@@ -10,35 +11,43 @@ exports.handler = async (event) => {
         return api.successResponse({});
     }
 
-    let method = event.pathParameters.proxy;
-    let body = null;
-    if (event.body) {
-        body = JSON.parse(event.body);
-    }
-    let authorizedUser = event.requestContext.authorizer.claims["cognito:username"];
-    let email = event.requestContext.authorizer.claims.email;
-
-    let responsePromise = (async function(method) {
-        switch(method) {
-            case 'create':
-                await validate.create(body, authorizedUser, email);
-                console.log("Create validation passed");
-                return api.create(body, email);
-            case 'read':
-                await validate.read(body);
-                return api.read(body, email);
-            case 'update':
-                await validate.update(body);
-                return api.update(body, email);
-            case 'delete':
-                await validate.delete(body);
-                return api.delete(body, email);
-            case 'list':
-                return api.list(email);
-            default:
-                return Promise.reject({message: "Unrecognized method name ".concat(method)});
+    // Pass CodePipeline events to clean.js, others to api.js
+    let responsePromise;
+    if (event['CodePipeline.job']){
+        responsePromise = cleanup.postPipelineCleanup(event['CodePipeline.job'].data);
+    } else if (event.httpMethod){
+        let method = event.pathParameters.proxy;
+        let body = null;
+        if (event.body) {
+            body = JSON.parse(event.body);
         }
-    })(method);
+        let authorizedUser = event.requestContext.authorizer.claims["cognito:username"];
+        let email = event.requestContext.authorizer.claims.email;
+
+        responsePromise = (async function(method) {
+            switch(method) {
+                case 'create':
+                    await validate.create(body, authorizedUser, email);
+                    console.log("Create validation passed");
+                    return api.create(body, email);
+                case 'read':
+                    await validate.read(body);
+                    return api.read(body, email);
+                case 'update':
+                    await validate.update(body);
+                    return api.update(body, email);
+                case 'delete':
+                    await validate.delete(body);
+                    return api.delete(body, email);
+                case 'list':
+                    return api.list(email);
+                default:
+                    return Promise.reject({message: "Unrecognized method name ".concat(method)});
+            }
+        })(method);
+    } else {
+        responsePromise = Promise.reject(`Lambda received an event it did not know how to parse: ${JSON.stringify(event, undefined, 2)}`)
+    }
 
     let response = null;
     try {
